+++ src/ArduinoJson/Document/JsonDocument.hpp
@@ -14,7 +14,8 @@
 
 namespace ARDUINOJSON_NAMESPACE {
 
-class JsonDocument : public Visitable {
+class JsonDocument : public Visitable,
+                     public VariantOperators<const JsonDocument&> {
  public:
   template <typename TVisitor>
   typename TVisitor::result_type accept(TVisitor& visitor) const {
@@ -295,16 +296,12 @@ class JsonDocument : public Visitable {
     _data.remove(adaptString(key));
   }
 
+  FORCE_INLINE operator VariantRef() {
+    return getVariant();
+  }
+
   FORCE_INLINE operator VariantConstRef() const {
-    return VariantConstRef(&_data);
-  }
-
-  bool operator==(VariantConstRef rhs) const {
-    return getVariant() == rhs;
-  }
-
-  bool operator!=(VariantConstRef rhs) const {
-    return getVariant() != rhs;
+    return getVariant();
   }
 
  protected: